+++ convoy/fleet.py
@@ -64,7 +64,7 @@ _REQUEST_CHUNK_SIZE = 4194304
 _ROOT_PATH = pathlib.Path(__file__).resolve().parent.parent
 _RESOURCES_PATH = None
 _NVIDIA_DRIVER = {
-    'compute': {
+    'compute_cc37': {
         'url': (
             'http://us.download.nvidia.com/tesla/'
             '384.111/NVIDIA-Linux-x86_64-384.111.run'
@@ -72,9 +72,19 @@ _NVIDIA_DRIVER = {
         'sha256': (
             'bd8af7654ccb224c37e74c8e81477a42f63fa9f2360b1b1ec6ae00b03ae21054'
         ),
-        'target': 'nvidia-driver.run'
+        'target': 'nvidia-driver_cc37.run'
     },
-    'visualization': {
+    'compute_cc6-7': {
+        'url': (
+            'http://us.download.nvidia.com/tesla/'
+            '390.30/NVIDIA-Linux-x86_64-390.30.run'
+        ),
+        'sha256': (
+            'ca4cb41a3c472a86b9c31784f07c6d2186539b12d2cdffd11c2152013b687e83'
+        ),
+        'target': 'nvidia-driver_cc6-7.run'
+    },
+    'viz_cc52': {
         'url': 'https://go.microsoft.com/fwlink/?linkid=849941',
         'sha256': (
             'ca3fd5f5e9156ad3d983b2032bde3c009dca73400f2753f9b475825f4670a854'